+++ lib/video/video_with_metada/video_with_metadata.g.dart
@@ -17,28 +17,43 @@ VideoWithMetadata _$VideoWithMetadataFromJson(Map<String, dynamic> json) =>
       appUrl: json['appUrl'] as String?,
       name: json['name'] as String?,
       description: json['description'] as String?,
+      partnerId: json['partnerId'] as String?,
       order: (json['order'] as num?)?.toInt(),
       active: json['active'] as bool?,
       createAt: json['createAt'] == null
           ? null
           : DateTime.parse(json['createAt'] as String),
-      partnerId: json['partnerId'] as String?,
+      streamingData: json['streamingData'] == null
+          ? null
+          : StreamingData.fromJson(
+              json['streamingData'] as Map<String, dynamic>),
+      isFeaturedContent: json['isFeaturedContent'] as bool?,
+      partnerName: json['partnerName'] as String?,
+      shared: (json['shared'] as num?)?.toInt(),
+      followingCO: json['followingCO'] as bool?,
+      likeThisVideo: json['likeThisVideo'] as bool?,
       isSubscriberOnly: json['isSubscriberOnly'] as bool?,
       hasAccess: json['hasAccess'] as bool?,
     );
 
 Map<String, dynamic> _$VideoWithMetadataToJson(VideoWithMetadata instance) =>
     <String, dynamic>{
-      'videoMetadata': instance.videoMetadata,
+      'videoMetadata': instance.videoMetadata?.toJson(),
       'id': instance.id,
       'videoUrl': instance.videoUrl,
       'appUrl': instance.appUrl,
       'name': instance.name,
       'description': instance.description,
+      'partnerId': instance.partnerId,
       'order': instance.order,
       'active': instance.active,
       'createAt': instance.createAt?.toIso8601String(),
-      'partnerId': instance.partnerId,
+      'streamingData': instance.streamingData?.toJson(),
+      'isFeaturedContent': instance.isFeaturedContent,
+      'partnerName': instance.partnerName,
+      'shared': instance.shared,
+      'followingCO': instance.followingCO,
+      'likeThisVideo': instance.likeThisVideo,
       'isSubscriberOnly': instance.isSubscriberOnly,
       'hasAccess': instance.hasAccess,
     };